+++ src/service-workers/resources-only.js
@@ -26,7 +26,6 @@ self.addEventListener("fetch", event => {
 
                 return fetch(event.request, {
                     redirect: "follow",
-                    credentials: "include",
                 }).then(response => {
                     if (!response || response.status !== 200 || response.type !== "basic" || response.headers.get("PWA-Cache") === "no-cache" || response.redirected) {
                         return response;